backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
ultralytics==8.2.82
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
gunicorn==22.0.0
